feat: parse more of the header flags
Now parses the opcode and returncode/status fields of the header.

Required bit-level parsing, so used nom's builtin support for that.
Potentially could have been easier to implement that manually with
bitmasks because the nom stuff required some fairly gnarly type
annotations to get it working, but hopefully it's clearer as a result
now.